add: tabbar에 헬스킷 권한 요청 분리
