+++ Booster/Booster/ViewControllers/Home/HomeViewController.swift
@@ -24,25 +24,7 @@ final class HomeViewController: UIViewController {
 
     // MARK: - Functions
     func configure() {
-        configureHealthKit()
         bindHomeViewModel()
-    }
-
-    private func configureHealthKit() {
-        guard let activeEnergyBurned = HKSampleType.quantityType(forIdentifier: .activeEnergyBurned),
-              let distanceWalkingRunning = HKSampleType.quantityType(forIdentifier: .distanceWalkingRunning),
-              let stepCount = HKSampleType.quantityType(forIdentifier: .stepCount)
-        else { return }
-
-        let shareTypes = Set([activeEnergyBurned, distanceWalkingRunning, stepCount])
-        let readTypes = Set([activeEnergyBurned, distanceWalkingRunning, stepCount])
-
-        HealthKitManager.shared.requestAuthorization(shareTypes: shareTypes, readTypes: readTypes)
-            .subscribe { [weak self] requestResult in
-                if case .success = requestResult {
-                    self?.viewModel.fetchQueries()
-                }
-            }.disposed(by: disposeBag)
     }
 
     private func bindHomeViewModel() {
